pyproject.toml
[tool.poetry]
name = "plamo-100b"
version = "0.1.0"
description = ""
authors = ["naoki <u429561d@ecs.osaka-u.ac.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.40.2"
python-dotenv = "^1.0.1"
streamlit = "^1.37.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
